Add support for more complex gestures
This adds support for more complex gestures. These gestures can be
added to the ftalk.h file, currently a limited subset of gestures are
in there.  The time delay between the last input even and the pruning
of gestures is currently set to 1000 milliseconds, this will need to
be tuned.

Signed-off-by: Brian Pomerantz <bapper@gmail.com>